substitui int por double

+++ funcoes.c
@@ -1,23 +1,23 @@
 #include <stdio.h>
 #include "funcoes.h"
 
-int soma(int a, int b) {
+int soma(double a, double b) {
     return a + b;
 }
 
-int subtracao(int a, int b) {
+int subtracao(double a, double b) {
 	return a - b;
 }
 
-int multiplicacao(int a, int b) {
+int multiplicacao(double a, double b) {
 	return a*b;
 }
 
-double divisao(int a, int b) {
+double divisao(double a, double b) {
     if (b == 0) {
         printf("Erro: divisão por zero!\n");
         return 0;  // ou outra forma de tratar
     }
-    return (double)a / b;
+    return a / b;
 }
